+++ src/command_modules/azure-cli-redis/HISTORY.rst
@@ -3,6 +3,8 @@
 Release History
 ===============
 
+0.2.0 (2017-04-17)
+++++++++++++++++++++
 * Adding update command which also adds the ability to scale for redis cache
 * Deprecates the 'update-settings' command.